requirements.txt
# requirements.txt
python-telegram-bot==20.8
requests>=2.31.0
python-dotenv>=1.0.0

# Discord
discord.py>=2.0
PyNaCl>=1.5.0

# Gmail
google-api-python-client>=2.0
google-auth>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.1

# AI (local)
ollama>=0.3.0